split session module into separate modules per class defs
